+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### To Release v0.6.0
 
+* [project] remove directory
+* [project] add/remove file
 * [vc] c+k will push commit editor to front
 * [project] ignore some useless files/directories
 * [vc:fix] handle AM/UU/MM prefix as changes